+++ debug_toolbar/panels/redirects.py
@@ -1,3 +1,4 @@
+import warnings
 from inspect import iscoroutine
 
 from django.template.response import SimpleTemplateResponse
@@ -16,6 +17,17 @@ class RedirectsPanel(Panel):
     is_async = True
 
     nav_title = _("Intercept redirects")
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        warnings.warn(
+            "The RedirectsPanel is deprecated and will be removed in a future version. "
+            "The HistoryPanel now provides the ability to view toolbar data for redirected requests. "
+            "If you still have a use case for this panel, please comment on "
+            "https://github.com/django-commons/django-debug-toolbar/issues/2216",
+            DeprecationWarning,
+            stacklevel=2,
+        )
 
     def _process_response(self, response):
         """